Make survey response answer have its own endpoint file
