+++ packages/lan/app/routes/apiv1/index.js
@@ -35,6 +35,7 @@ import { scheduledVaccine } from './scheduledVaccine';
 import { suggestions } from './suggestions';
 import { survey } from './survey';
 import { surveyResponse } from './surveyResponse';
+import { surveyResponseAnswer } from './surveyResponseAnswer';
 import { sync } from './sync';
 import { syncHealth } from './syncHealth';
 import { triage } from './triage';
@@ -82,6 +83,7 @@ patientDataRoutes.use('/patientIssue', patientIssue);
 patientDataRoutes.use('/procedure', procedure);
 patientDataRoutes.use('/referral', referral);
 patientDataRoutes.use('/surveyResponse', surveyResponse);
+patientDataRoutes.use('/surveyResponseAnswer', surveyResponseAnswer);
 patientDataRoutes.use('/triage', triage);
 patientDataRoutes.use('/vitals', vitals);
